Move test_fake_scene to test_utils
Move the function that tests make_fake_scene from testing/utils.py to
testing/test_utils.py, so that it should be executed by default as part
of the testing suite, including on Travis.
