+++ satpy/tests/test_utils.py
@@ -248,3 +248,40 @@ class TestUtils(unittest.TestCase):
         dataset.attrs.pop('orbital_parameters')
         lon, lat, alt = get_satpos(dataset)
         self.assertTupleEqual((lon, lat, alt), (-1, -2, -3))
+
+
+def test_make_fake_scene():
+    """Test the make_fake_scene utility.
+
+    Although the make_fake_scene utility is for internal testing
+    purposes, it has grown sufficiently complex that it needs its own
+    testing.
+    """
+    import numpy as np
+    import dask.array as da
+    import xarray as xr
+    from satpy.tests.utils import make_fake_scene
+
+    assert make_fake_scene({}).keys() == []
+    sc = make_fake_scene({
+        "six": np.arange(25).reshape(5, 5)})
+    assert len(sc.keys()) == 1
+    assert sc.keys().pop().name == "six"
+    assert sc["six"].attrs["area"].shape == (5, 5)
+    sc = make_fake_scene({
+        "seven": np.arange(3*7).reshape(3, 7),
+        "eight": np.arange(3*8).reshape(3, 8)},
+        daskify=True,
+        area=False,
+        common_attrs={"repetency": "fourteen hundred per centimetre"})
+    assert "area" not in sc["seven"].attrs.keys()
+    assert (sc["seven"].attrs["repetency"] == sc["eight"].attrs["repetency"] ==
+            "fourteen hundred per centimetre")
+    assert isinstance(sc["seven"].data, da.Array)
+    sc = make_fake_scene({
+        "nine": xr.DataArray(
+            np.arange(2*9).reshape(2, 9),
+            dims=("y", "x"),
+            attrs={"please": "preserve", "answer": 42})},
+        common_attrs={"bad words": "semprini bahnhof veerooster winterbanden"})
+    assert sc["nine"].attrs.keys() >= {"please", "answer", "bad words", "area"}
